[FAKE] moved StubConsole into b3.fake from b3.setup
