+++ b3/fake.py
@@ -40,6 +40,7 @@
 # 1.16 - 2015/01/29 - make use of the new b3.config.MainConfig class
 # 1.17 - 2015/02/07 - update write method to accept a socketTimeout=None parameter
 # 1.18 - 2015/03/19 - removed deprecated usage of dict.has_key (us 'in dict' instead)
+# 1.19 - 2015/05/26 - moved StubConsole from b3.setup
 
 """
 This module make plugin testing simple. It provides you
@@ -47,7 +48,7 @@ with fakeConsole and joe which can be used to say commands
 as if it where a player.
 """
 
-__version__ = '1.18'
+__version__ = '1.19'
 
 import b3.events
 import b3.output
@@ -69,7 +70,9 @@ from b3.storage.sqlite import SqliteStorage
 from sys import stdout
 
 class FakeConsole(b3.parser.Parser):
-
+    """
+    Console implementation to be used with automated tests.
+    """
     Events = b3.events.eventManager = b3.events.Events()
     screen = stdout
     noVerbose = False
@@ -127,8 +130,7 @@ class FakeConsole(b3.parser.Parser):
         """
         NO QUEUE, NO THREAD for faking speed up
         """
-        if event.type == self.getEventID('EVT_EXIT') or \
-                event.type == self.getEventID('EVT_STOP'):
+        if event.type == self.getEventID('EVT_EXIT') or event.type == self.getEventID('EVT_STOP'):
             self.working = False
 
         nomore = False
@@ -275,7 +277,9 @@ class FakeConsole(b3.parser.Parser):
 
 
 class FakeClient(b3.clients.Client):
-
+    """
+    Client object implementation to be used in automated tests.
+    """
     console = None
 
     def __init__(self, console, **kwargs):
@@ -395,6 +399,47 @@ class FakeClient(b3.clients.Client):
         self.console.queueEvent(b3.events.Event(type, data, self, target))
 
 #####################################################################################
+
+
+class StubConsole(object):
+    """
+    Console implementation used when dealing with the Storage module while updating B3 database.
+    """
+
+    screen = sys.stdout
+
+    @staticmethod
+    def bot(self, msg, *args, **kwargs):
+        pass
+
+    @staticmethod
+    def info(self, msg, *args, **kwargs):
+        pass
+
+    @staticmethod
+    def debug(self, msg, *args, **kwargs):
+        pass
+
+    @staticmethod
+    def error(self, msg, *args, **kwargs):
+        pass
+
+    @staticmethod
+    def warning(self, msg, *args, **kwargs):
+        pass
+
+    @staticmethod
+    def verbose(self, msg, *args, **kwargs):
+        pass
+
+    @staticmethod
+    def verbose2(self, msg, *args, **kwargs):
+        pass
+
+    @staticmethod
+    def critical(msg, *args, **kwargs):
+        raise SystemExit(msg)
+
 
 print "creating fakeConsole with @b3/conf/b3.distribution.ini"
 fakeConsole = FakeConsole('@b3/conf/b3.distribution.ini')
